+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-27 20:19:47
-// Commit: test(state): add unit tests for state business logic
+// Updated at: 2026-05-28 09:50:45
+// Commit: feat(state): add bi-directional event bus handler for state
 
 class stateComponentState {
-  final String commitRef = '1525';
-  final String lastMessage = 'test(state): add unit tests for state business logic';
-  final DateTime timestamp = DateTime.parse('2026-05-27T20:19:47+05:00');
+  final String commitRef = '1542';
+  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
+  final DateTime timestamp = DateTime.parse('2026-05-28T09:50:45+05:00');
 }